requirements.txt
asgiref==3.8.1
cloudinary==1.36.0
crispy-bootstrap5==0.7
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.11
django-admin-rangefilter==0.12.4
django-crispy-forms==2.1
django-summernote==0.8.20.0
gunicorn==20.1.0
psycopg2==2.9.9
sqlparse==0.4.4
urllib3==1.26.18
whitenoise==5.3.0
